microservices/simulator/requirements.txt
flask>=2.0.0
gunicorn>=20.1.0
kafka-python>=2.0.2
flask-cors
flask-sock
cryptography